Search for gh executable on different platforms
On mac and linux, launching from the icon does not create a shell,
so the custom install paths are not available.  For this, search
in the default homebrew / linuxbrew locations and use gh if found.

For windows the path is set when lauching from the start menu, but
the path to the executable is not easily available, so we search for it
by iterating through the elements of the PATH variable.

Once the executable is found, the result is stored in the settings for
future use.

The user can also set a different gh executable via the advanced
settings of the MorphoDepot module.