Use pixi to install git and gh
We pull in our own copies of git and gh so that users don't need
to install them manually.

Note that because git is GPL, we can't bundle the executable with
our extension and need the user to install it themselves.

Pixi uses conda-forge packages under the hood, so this solution should
be robust.

Fixes #2