Refactor git/gh to require system install
Addresses #24

For now, pixi support is left in place in case we decide to
use this method or a similar one to make the installation process
easier for end-users.

The current code disables all modules if working versions of git
and gh are not found on the system.  The user is prompted to
refer to the instructions to install them for their platform.